removed directory auto-creation when not needed
