+++ src/main/java/config/Config.java
@@ -4,9 +4,6 @@ import com.intellij.ide.util.PropertiesComponent;
 
 import java.awt.*;
 import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 
 public class Config {
     public static String DEFAULT_PATH_SUFFIX = "/.IntelliJPlugin/DebuggerVariableSaveLoader/";
@@ -27,12 +24,6 @@ public class Config {
         String[] values = PropertiesComponent.getInstance().getValues(DIMENSION_KEY);
         if (values != null) {
             tableDimension = new Dimension(Integer.parseInt(values[0]), Integer.parseInt(values[1]));
-        }
-
-        try {
-            Files.createDirectories(Paths.get(DEFAULT_PATH_ABSOLUTE));
-        } catch (IOException e) {
-            throw new RuntimeException("Cannot create working directory", e);
         }
     }
 }
