+++ src/main/resources/META-INF/plugin.xml
@@ -56,7 +56,7 @@
             <add-to-group group-id="XDebugger.ValueGroup" anchor="after" relative-to-action="XDebugger.CopyName"/>
         </action>
         <action class="action.CopyValueAction" id="CopyValueAction"
-                        text="Copy Value as Java Code" description="Copy value as Java code.">
+                        text="Extract Value..." description="Extract value as Java or JSON code.">
             <add-to-group group-id="XDebugger.ValueGroup" anchor="after" relative-to-action="XDebugger.CopyValue"/>
             <keyboard-shortcut keymap="$default" first-keystroke="meta alt C"/>
         </action>
